Added new function list_items_to_english.
This function builds an english list out of python list, complete with
oxford comma if requested.  Single item lists are just returned as the
list, two item lists are simply separated by the word and, and list with
three or more items are comma separated lists with an optional oxford
comma and the last word separated by the word and.
['one'] -> one
['one', 'two'] -> one and two
['one', 'two', 'three'] -> one, two, and three